Added user interaction
Once the machines are created the user now has the ability to list what
was created, the remote IP, the port used for the socks tunnel, and if
there is currently an active tunnel.

The user also has the ability to disconnect an individual tunnel. This
will remove the port from the list option, set the active ssh session to
false and stop the ssh -D process.

The user will also have the option to add a connection to a host if it
currently does not have one. At this point we don't allow for multiple
socks proxies per IP.